+++ src/main/java/com/inz/inz/adapter/ReportAdapter.java
@@ -3,6 +3,7 @@ package com.inz.inz.adapter;
 import com.inz.inz.ExcpetionHandler.DbException;
 import com.inz.inz.ExcpetionHandler.EnumExcpetion;
 import com.inz.inz.entity.ReportEntity;
+import com.inz.inz.resoruce.ReportResource;
 import com.inz.inz.resoruce.ReportResourcePost;
 
 import javax.servlet.http.HttpServletRequest;
@@ -11,4 +12,5 @@ public interface ReportAdapter {
 
     ReportEntity createReport(HttpServletRequest request, ReportResourcePost reportResourcePost) throws DbException, EnumExcpetion;
 
+    ReportResource getReport(Long id) throws DbException;
 }