+++ src/main/java/com/inz/inz/controller/ReportController.java
@@ -3,14 +3,12 @@ package com.inz.inz.controller;
 import com.inz.inz.ExcpetionHandler.DbException;
 import com.inz.inz.ExcpetionHandler.EnumExcpetion;
 import com.inz.inz.adapter.adapterImpl.ReportAdapterImpl;
+import com.inz.inz.resoruce.ReportResource;
 import com.inz.inz.resoruce.ReportResourcePost;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
@@ -22,6 +20,13 @@ public class ReportController {
 
     @Autowired
     ReportAdapterImpl reportAdapter;
+
+    @GetMapping(value = "/getReport/{id}", produces = MEDIA_TYPE )
+    public ResponseEntity<ReportResource> addReport(@PathVariable Long id ) throws DbException {
+       ReportResource reportResource= reportAdapter.getReport(id);
+        return new ResponseEntity<>(reportResource,HttpStatus.CREATED);
+    }
+
     @PostMapping(value = "/addReport", consumes = MEDIA_TYPE, produces = MEDIA_TYPE )
     public ResponseEntity<?> addReport(@RequestBody @Valid ReportResourcePost reportResourcePost, HttpServletRequest request) throws DbException, EnumExcpetion {
         reportAdapter.createReport(request,reportResourcePost);