+++ src/main/java/com/inz/inz/mapper/ReportMapper.java
@@ -2,12 +2,14 @@ package com.inz.inz.mapper;
 
 import com.inz.inz.ExcpetionHandler.DbException;
 import com.inz.inz.ExcpetionHandler.EnumExcpetion;
+import com.inz.inz.ExcpetionHandler.ErrorSpecifcation;
 import com.inz.inz.ExcpetionHandler.Field;
 import com.inz.inz.entity.CityEntity;
 import com.inz.inz.entity.ReportEntity;
 import com.inz.inz.entity.enums.ReportType;
 import com.inz.inz.repository.CityEntityRepository;
 import com.inz.inz.resoruce.ReportLight;
+import com.inz.inz.resoruce.ReportResource;
 import com.inz.inz.resoruce.ReportResourcePost;
 import org.mapstruct.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,23 +17,39 @@ import org.springframework.beans.factory.annotation.Autowired;
 import java.util.Date;
 import java.util.Optional;
 
-@Mapper(componentModel="spring")
+@Mapper(componentModel = "spring")
 public abstract class ReportMapper {
 
     @Autowired
     CityEntityRepository cityEntityRepository;
 
 
+
+
     @Mappings({
-            @Mapping(target = "id",source = "res.id"),
+            @Mapping(target = "id", source = "res.id"),
             @Mapping(target = "longitude", source = "res.longitude"),
             @Mapping(target = "latitude", source = "res.latitude"),
             @Mapping(target = "description", source = "res.description"),
             @Mapping(target = "photo", source = "res.photo"),
-            @Mapping(target = "video", source = "res.video"),
-
+            @Mapping(target = "video", source = "res.video")
     })
-    public  abstract ReportLight mapToReportLigth(ReportEntity res);
+    public abstract ReportLight mapToReportLigth(ReportEntity res);
+
+    @Mappings({
+            @Mapping(target = "id", source = "res.id"),
+            @Mapping(target = "longitude", source = "res.longitude"),
+            @Mapping(target = "latitude", source = "res.latitude"),
+            @Mapping(target = "description", source = "res.description"),
+            @Mapping(target = "photo", source = "res.photo"),
+            @Mapping(target = "video", source = "res.video"),
+            @Mapping(target = "reportDate",expression = "java(res.getDateReport().toString())"),
+            @Mapping(target = "userId",expression = "java(res.getUser().getId())"),
+            @Mapping(target = "reportRating",ignore = true),
+            @Mapping(target = "mark", expression = "java( res.getReportRating().getQuantity()!=0?res.getReportRating().getMarks()/res.getReportRating().getQuantity():0.0)"),
+            @Mapping(target = "isActive", expression = "java( res.getReportRating().getNotActiveCounter()<10?true:false)")
+    })
+    public abstract ReportResource mapToReport(ReportEntity res);
 
     @Mappings({
             @Mapping(target = "longitude", source = "res.longitude"),
@@ -39,7 +57,7 @@ public abstract class ReportMapper {
             @Mapping(target = "description", source = "res.description"),
             @Mapping(target = "photo", source = "res.photo"),
             @Mapping(target = "video", source = "res.video"),
-            @Mapping(target = "reportType",ignore = true)
+            @Mapping(target = "reportType", ignore = true)
     })
     public abstract ReportEntity mapToEntity(ReportResourcePost res) throws DbException, EnumExcpetion;
 
@@ -48,9 +66,9 @@ public abstract class ReportMapper {
         Optional<CityEntity> cityEntity = cityEntityRepository.findByName(res.getCityName());
         if (!cityEntity.isPresent()) {
             Field field = new Field();
-            field.setField(null);
-            field.setDetails("City doesn't exist");
-            throw new DbException("Error at getting info from cityEntity ", "D3", field);
+            field.setField("cityName");
+            field.setDetails("City " + res.getCityName());
+            throw new DbException(ErrorSpecifcation.RESURCENOTEXIST.getDetails(), ErrorSpecifcation.RESURCENOTEXIST.getCode(), field);
         } else {
             entity.setCity(cityEntity.get());
         }
@@ -59,7 +77,9 @@ public abstract class ReportMapper {
         try {
             entity.setReportType(ReportType.valueOf(res.getReportType()));
         } catch (IllegalArgumentException ex) {
-            throw new EnumExcpetion("ReportTypeNotExist", "E1", new Field());
+            throw new EnumExcpetion(ErrorSpecifcation.RESURCENOTEXIST.getDetails() + res.getReportType(), ErrorSpecifcation.RESURCENOTEXIST.getCode(), new Field());
         }
     }
+
+
 }