+++ changelog.md
@@ -1,3 +1,11 @@
+
+##[0.0.0.2]
+### changes
+- add enum errospeciftiation 
+- correct throwing excpetion 
+- add getting active/notactive rpeorts in city 
+
+
 ## [0.0.0.1]
 ### changes
 - add jwt token auth